Assignment2/requirements.txt
torch
torchvision
matplotlib
numpy<2
pillow
